requirements.txt
alabaster==0.7.16
annotated-types==0.6.0
anyio==4.3.0
Babel==2.14.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.0
cryptography==43.0.1
cycler==0.11.0
docutils==0.19
flatbuffers==23.5.26
fonttools==4.40.0
h11==0.14.0
httptools==0.6.1
humanfriendly==10.0
idna==3.6
imagesize==1.4.1
Jinja2==3.1.3
kiwisolver==1.4.4
MarkupSafe==2.1.5
matplotlib==3.7.1
mpmath==1.3.0
none==0.1.1
numpy==1.25.0
onnx==1.14.0
onnxruntime==1.15.1
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
protobuf==4.23.3
pycocotools==2.0.6
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
rfc3986==2.0.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sphinxcontrib-websupport==1.2.7
starlette==0.37.2
sympy==1.12
typing_extensions==4.11.0
urllib3==2.2.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
